Added more ClickUp functions
todo: finish CREATE_TASK, make find_lists better
probably ask first for value of folderless, then have the bot send
another message asking you to fill in the value for either the folder or
space ID instead, since idk how to ask discord to force a value for
options in order for the list to stay the same and the data.options[x]
call to work if a user accidentally fills in both "folder_id" and "space_id". But it should still work as is naman.